Convert bench to a free-standing lib
