+++ .gitignore
@@ -1,3 +1,4 @@
 *.pdf
 *.pprof
 *.test
+bench/bench
