+++ bench/bench.go
@@ -1,4 +1,4 @@
-package bench
+package main
 
 import (
 	"encoding/hex"
@@ -11,24 +11,18 @@ import (
 	"time"
 )
 
-func init() {
-	log.SetFlags(log.Lshortfile)
-}
-
 var (
 	dbPath      = "data"
 	benchBucket = []byte("bench_bkt")
 )
 
-func rmDB(dbPath string) error {
-	return os.RemoveAll(dbPath)
-}
-
+// Value block
 var (
 	SmallVal = []byte(strings.Repeat("ABCD", 2*KB)) // 8KB
 	LargeVal = []byte(strings.Repeat("ABCD", 2*MB)) // 8MB
 )
 
+// Sizing reference block
 const (
 	KB = 1024
 	MB = 1024 * 1024
@@ -37,15 +31,40 @@ const (
 
 var take int
 
+var clock int
+
+var src = rand.New(rand.NewSource(1))
+
+func main() {
+	//	batchSize := int64(10000)
+	//	bench(100, 100, 3)
+	//bench(500, 100, 3)
+	bench(5000, 500, 3)
+
+	os.Exit(0)
+}
+
+func bench(benchCount, batchSize, iterations int64) {
+	benchFunc(whiskeyPut, benchCount, batchSize, iterations)
+	benchFunc(boltPut, benchCount, batchSize, iterations)
+	benchFunc(lmdbPut, benchCount, batchSize, iterations)
+}
+
+func benchFunc(fn benchFn, benchCount, batchSize, iterations int64) {
+	for i := int64(0); i < iterations; i++ {
+		fn(benchCount, batchSize, 2*GB, SmallVal)
+	}
+}
+
+func rmDB(dbPath string) error {
+	return os.RemoveAll(dbPath)
+}
+
 func checkErr(err error) {
 	if err != nil {
 		log.Fatalln("[ERR]", err)
 	}
 }
-
-var clock int
-
-var src = rand.New(rand.NewSource(1))
 
 // sortedKey returns key of (9 bytes of clock + 7 (or 14 chars) random bytes).
 func sortedKey() []byte {
@@ -79,3 +98,5 @@ func mem() {
 	runtime.ReadMemStats(&stats)
 	log.Printf("mem: heap %dMB/%dMB, alloc total %dMB\n", stats.HeapInuse/MB, stats.HeapAlloc/MB, stats.TotalAlloc/MB)
 }
+
+type benchFn = func(benchCount, batchSize, envSize int64, val []byte)
